+++ src/pages/Approvals/AddNewApproval.tsx
@@ -1,34 +1,17 @@
-import { useForm, useFieldArray, Controller } from "react-hook-form";
+import { useForm } from "react-hook-form";
 import DatePicker from "react-datepicker";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faClose } from "@fortawesome/free-solid-svg-icons";
+import { useHistory } from "react-router-dom";
 
 import { LabeledInput } from "components/Inputs/LabeledInput";
 import { Form } from "components/Inputs/Form";
 import { PersonSelect } from "components/PersonSelect";
-import {
-  ApprovalItemType,
-  PartialApproval,
-  PartialApprovalItem,
-} from "apiClient/approvals";
-import { Select } from "components/Select";
-import { GearTypeSelect } from "components/GearTypeSelect";
-import { GearItemSelect } from "components/GearItemSelect";
-import styled from "styled-components";
-import { useHistory } from "react-router-dom";
 
-type FormValues = PartialApproval;
-
-const defaultItem: PartialApprovalItem = {
-  type: ApprovalItemType.gearType,
-  item: {
-    quantity: 1,
-    gearType: undefined,
-  },
-};
+import { ApprovalItemsPicker, defaultItem } from "./ApprovalItemsPicker";
+import { FormValues } from "./types";
 
 // TODO: Validate dates are in the right order
 // TODO: Handle submit
+// TODO: Simplify with LabeledControlledInput?
 
 export function AddNewApproval() {
   const history = useHistory();
@@ -37,19 +20,10 @@ export function AddNewApproval() {
       items: [defaultItem],
     },
   });
-  const {
-    fields: itemFields,
-    append,
-    remove,
-  } = useFieldArray({
-    control: formObject.control, // control props comes from useForm (optional: if you are using FormContext)
-    name: "items", // unique name for your Field Array
-  });
 
   const onSubmit = (values: FormValues) => {
     console.log({ values });
   };
-  const items = formObject.watch("items");
   return (
     <div className="row">
       <div className="col-lg-8">
@@ -61,7 +35,6 @@ export function AddNewApproval() {
             renderComponent={({ value, onChange, onBlur, invalid }: any) => {
               return (
                 <PersonSelect
-                  className={`w-100 flex-grow-1 ${invalid ? "is-invalid" : ""}`}
                   value={value}
                   onChange={onChange}
                   invalid={invalid}
@@ -108,106 +81,7 @@ export function AddNewApproval() {
               required: true,
             }}
           />
-          <fieldset>
-            Items approved:
-            {itemFields.map((field, index) => {
-              return (
-                <StyledItem key={field.id} className="p-2 mb-3 mt-3">
-                  <div className="d-flex justify-content-between mb-3">
-                    <strong>Item #{index + 1}</strong>
-                    {items.length > 1 && (
-                      <button
-                        type="button"
-                        className="btn btn-outline-secondary"
-                        onClick={() => remove(index)}
-                        style={{
-                          borderColor: "#ced4da",
-                        }}
-                      >
-                        <FontAwesomeIcon icon={faClose} />
-                      </button>
-                    )}
-                  </div>
-                  <Controller
-                    control={formObject.control}
-                    name={`items.${index}.type`}
-                    rules={{ required: true }}
-                    render={({ field: { onChange, onBlur, value, ref } }) => {
-                      return (
-                        <ApprovalTypePicker
-                          value={value}
-                          onChange={(v) => {
-                            formObject.resetField(`items.${index}`);
-                            onChange(v);
-                          }}
-                          onBlur={onBlur}
-                        />
-                      );
-                    }}
-                  />
-                  {items[index].type === ApprovalItemType.gearType ? (
-                    <>
-                      <LabeledInput
-                        title="Type:"
-                        name={`items.${index}.item.gearType`}
-                        renderComponent={({
-                          value,
-                          onChange,
-                          invalid,
-                        }: any) => {
-                          return (
-                            <GearTypeSelect
-                              value={value ?? null}
-                              onChange={onChange}
-                              invalid={invalid}
-                            />
-                          );
-                        }}
-                        options={{
-                          required: true,
-                        }}
-                      />
-                      <LabeledInput
-                        title="Quantity approved"
-                        type="number"
-                        name={`items.${index}.item.quantity`}
-                        step={1}
-                        options={{
-                          valueAsNumber: true,
-                        }}
-                      />
-                    </>
-                  ) : (
-                    <LabeledInput
-                      title="Item:"
-                      name={`items.${index}.item.gearItem`}
-                      renderComponent={({ value, onChange, invalid }: any) => {
-                        return (
-                          <GearItemSelect
-                            value={value?.id ?? null}
-                            onChange={onChange}
-                            invalid={invalid}
-                          />
-                        );
-                      }}
-                      options={{
-                        required: true,
-                      }}
-                    />
-                  )}
-                </StyledItem>
-              );
-            })}
-            <div className="mt-2 d-flex justify-content-end">
-              <button
-                className="btn btn-outline-primary"
-                onClick={() => append(defaultItem)}
-                type="button"
-              >
-                Add item
-              </button>
-            </div>
-          </fieldset>
+          <ApprovalItemsPicker />
 
           <LabeledInput title="Note:" as="textarea" name="note" />
 
@@ -228,41 +102,3 @@ export function AddNewApproval() {
     </div>
   );
 }
-
-const approvalTypeOptions = [
-  {
-    value: ApprovalItemType.gearType,
-    label: "Approve gear by type",
-  },
-  {
-    value: ApprovalItemType.specificItem,
-    label: "Approve a specific item",
-  },
-];
-
-function ApprovalTypePicker({
-  onChange,
-  value,
-  onBlur,
-}: {
-  onChange: (value: ApprovalItemType | undefined) => void;
-  onBlur?: () => void;
-  value: ApprovalItemType;
-}) {
-  const selectedOption = approvalTypeOptions.find((o) => o.value === value);
-  return (
-    <Select
-      onChange={(newType) => onChange(newType?.value)}
-      value={selectedOption}
-      options={approvalTypeOptions}
-      onBlur={onBlur}
-    />
-  );
-}
-
-const StyledItem = styled.div`
-  margin-left: 30px;
-  border: solid 1px #ced4da;
-  border-radius: 0.375rem;
-  background-color: #f8f8f8;
-`;
