+++ src/pages/Approvals/types.ts
@@ -1,0 +1,3 @@
+import type { PartialApproval } from "apiClient/approvals";
+
+export type FormValues = PartialApproval;
